refactored util function inside karpetner addon; increased npm version for release
