+++ lib/utils/index.ts
@@ -1,6 +1,5 @@
 export * from './addon-utils';
 export * from './context-utils';
-export * from './json-utils';
 export * from './namespace-utils';
 export * from './object-utils';
 export * from './sa-utils';
