+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "@aws-quickstart/ssp-amazon-eks",
-    "version": "0.12.5",
+    "version": "0.12.6",
     "license": "Apache-2.0",
     "main": "dist/index.js",
     "types": "dist/index.d.ts",
